feat: 코스 목록 DTO 추가
