+++ android/app/src/main/java/io/coursepick/coursepick/data/CoursesResponseItem.kt
@@ -1,0 +1,10 @@
+package io.coursepick.coursepick.data
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CoursesResponseItem(
+    val geometry: Geometry?,
+    val properties: Properties?,
+    val type: String?,
+)
