+++ android/app/src/main/java/io/coursepick/coursepick/data/Properties.kt
@@ -1,0 +1,10 @@
+package io.coursepick.coursepick.data
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Properties(
+    val distance: Double?,
+    val length: Double?,
+    val name: String?,
+)
